requirements.txt
pyyaml
scipy
keras
numpy
theano
